requirements.txt
django==4.0
mysql-connector-python==8.0.29
pymysql==1.0.2
pillow==9.1.1
django-countries==7.3.2
requests==2.27.1
djangorestframework==3.13.1
django-cors-headers==3.12.0
djangorestframework-simplejwt==5.1.0